+++ migrations/008_bathroom_to_integer.sql
@@ -1,5 +1,0 @@
--- Migration: Change bathroom column from TEXT[] to INTEGER[]
-
-ALTER TABLE subscriptions
-ALTER COLUMN bathroom TYPE INTEGER[]
-USING bathroom::INTEGER[];